move initializeGui and markPostion to ui manger
